- Experimental attempt to avoid var size issues with time since recently comp_timeb() gives back int64_t
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@10204 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ csctapi/ifd_smartreader.c
@@ -91,7 +91,7 @@ static int32_t smart_read(struct s_reader *reader, unsigned char *buff, uint32_t
 	struct sr_data *crdr_data = reader->crdr_data;
 	int32_t ret = 0;
 	uint32_t  total_read = 0;
-	int32_t gone = 0;
+	int64_t gone = 0;
 	struct timeb start, now;
 
 	cs_ftime(&start);
