+++ oscam-work.c
@@ -216,10 +216,10 @@ void *work_thread(void *ptr)
 
 			struct timeb actualtime;
 			cs_ftime(&actualtime);
-			int32_t gone = comp_timeb(&actualtime, &data->time);
+			int64_t gone = comp_timeb(&actualtime, &data->time);
 			if(data != &tmp_data && gone > (int) cfg.ctimeout+1000)
 			{
-				cs_debug_mask(D_TRACE, "dropping client data for %s time %d ms", username(cl), gone);
+				cs_debug_mask(D_TRACE, "dropping client data for %s time %"PRId64" ms", username(cl), gone);
 				__free_job_data(cl, data);
 				continue;
 			}
